==> notebooks/1_Extractor: Cost of Living - BeautifulSoup.ipynb <==
### *Cost of Living*  - Scrapeo  BeautifulSoup

A continuación, vamos a scrapearnos con BeautifulSoup los datos de coste de vida del periodo comprendido entre 2017/2021: 

import requests
import re
import pandas as pd
import numpy as np
from bs4 import BeautifulSoup

#### Año 2017 

Crearemos un diccionario donde iremos almacenando la información que screapeamos. Ya que los únicos datos que nos importa quedarnos son los del coste de vida genérico de cada país, a la hora de introducir la información en el diccionario, meteremos un `start stop step`. Esto podemos hacerlo gracias a que la información sigue una estructura fija. 

cost_living_2017= {"countries":[], 
                   "cost_living_index": [], 
                        }

url17= "https://www.numbeo.com/cost-of-living/rankings_by_country.jsp?title=2017"

res17= requests.get(url17)

sopa17 = BeautifulSoup(res17.content, 'html.parser') 
#sopa17

country17 = sopa17.find_all("td", {"class": "cityOrCountryInIndicesTable"})
#country17

for i in country17: 
    cost_living_2017["countries"].append(i.getText())

#cost_living_2017["countries"]

cli17 = sopa17.find_all("td", {"style":"text-align: right"})
#cli17

len(cli17)

for i in cli17[:726:6]: 
    cost_living_2017["cost_living_index"].append(i.getText())

costliving_2017= pd.DataFrame(cost_living_2017)
costliving_2017

costliving_2017.to_csv('costliving_2017.csv')

#### Año 2018 

cost_living_2018= {"countries":[], 
                   "cost_living_index": [], 
                        }

url18= "https://www.numbeo.com/cost-of-living/rankings_by_country.jsp?title=2018"

res18= requests.get(url18)

sopa18 = BeautifulSoup(res18.content, 'html.parser') 
#sopa18

country18 = sopa18.find_all("td", {"class": "cityOrCountryInIndicesTable"})
#country18

for i in country18: 
    cost_living_2018["countries"].append(i.getText())

#cost_living_2018["countries"]

cli18 = sopa18.find_all("td", {"style":"text-align: right"})
#cli18

len(cli18)

for i in cli18[:690:6]: 
    cost_living_2018["cost_living_index"].append(i.getText())

costliving_2018= pd.DataFrame(cost_living_2018)
costliving_2018

costliving_2018.to_csv('costliving_2018.csv')

#### Año 2019

cost_living_2019= {"countries":[], 
                   "cost_living_index": [], 
                        }

url19= "https://www.numbeo.com/cost-of-living/rankings_by_country.jsp?title=2019"

res19= requests.get(url19)

sopa19 = BeautifulSoup(res19.content, 'html.parser') 
#sopa19

country19 = sopa19.find_all("td", {"class": "cityOrCountryInIndicesTable"})
#country19

for i in country19: 
    cost_living_2019["countries"].append(i.getText())

#cost_living_2019["countries"]

cli19 = sopa19.find_all("td", {"style":"text-align: right"})
#cli19

len(cli19)

for i in cli19[:714:6]: 
    cost_living_2019["cost_living_index"].append(i.getText())

costliving_2019= pd.DataFrame(cost_living_2019)
costliving_2019

costliving_2019.to_csv('costliving_2019.csv')

#### Año 2020

cost_living_2020= {"countries":[], 
                   "cost_living_index": [], 
                        }

url20= "https://www.numbeo.com/cost-of-living/rankings_by_country.jsp?title=2020"

res20= requests.get(url20)

sopa20 = BeautifulSoup(res20.content, 'html.parser') 
#sopa20

country20 = sopa20.find_all("td", {"class": "cityOrCountryInIndicesTable"})
#country20

for i in country20: 
    cost_living_2020["countries"].append(i.getText())

#cost_living_2020["countries"]

cli20 = sopa20.find_all("td", {"style":"text-align: right"})
#cli20

len(cli20)

for i in cli20[:792:6]: 
    cost_living_2020["cost_living_index"].append(i.getText())

costliving_2020= pd.DataFrame(cost_living_2020)
costliving_2020

#### Año 2021 

cost_living_2021= {"countries":[], 
                   "cost_living_index": [], 
                        }

url21= "https://www.numbeo.com/cost-of-living/rankings_by_country.jsp?title=2021"

res21= requests.get(url21)

sopa21 = BeautifulSoup(res21.content, 'html.parser') 
#sopa21

country21 = sopa21.find_all("td", {"class": "cityOrCountryInIndicesTable"})
#country21

for i in country21: 
    cost_living_2021["countries"].append(i.getText())

costliving_2021= pd.DataFrame(cost_living_2020)
costliving_2021
