==> notebooks/2_Extractor: Natalidad  - Selenium .ipynb <==
# Most Polluted Countries - Web Scraping 

### · Antes de empezar... 

1. Para gestionar los Dataframes:

pip install plotly.express

import pandas as pd
import numpy as np

2. Para hacer web scraping con Selenium 

pip install selenium

pip install webdriver-manager

from selenium import webdriver
from webdriver_manager.chrome import ChromeDriverManager
from selenium.webdriver.common.keys import Keys
from selenium.webdriver.chrome.options import Options
from time import sleep

import warnings
warnings.filterwarnings('ignore')

from tqdm import tqdm

De cara a trabajar con Selenium

opciones= Options()
opciones.add_experimental_option('excludeSwitches', ['enable-automation'])

opciones.add_experimental_option('useAutomationExtension', False)
opciones.add_argument('--start-maximized') #empezar maximizado
opciones.add_argument('user.data-dir=selenium') #guarda las cookies
opciones.add_argument('--incognito')#incognito window

driver = webdriver.Chrome(ChromeDriverManager().install())

###  · Carga de CSVs

Empezamos cargando datos CSV procedentes de Kaggle con la información de los países y las ciudades más contaminadas desde 2017 hasta 2021. 

df_pcountries = pd.read_csv("datos/AIR QUALITY INDEX- top countries.csv", index_col = 0 )
df_pcountries.head(2)

df_pcities = pd.read_csv("datos/AIR QUALITY INDEX (by cities) - IQAir.csv", index_col = 0 )
df_pcities.head(2)

### · Web Scraping Selenium 

Nuestro objetivo es scrapear los datos de natalidad de distintos años/países del Expansión, para ello utilizaremos Selenium. Iniciamos estructurando como haríamos la extracción de uno de los años. En este caso el más inmediato, 2021:

# iniciamos el driver
driver = webdriver.Chrome(ChromeDriverManager().install())

## accedemos a la pagina web
driver.get("https://datosmacro.expansion.com/demografia/natalidad?anio=2021")
driver.implicitly_wait(10)

### aceptamos las cookies
driver.find_element("css selector", "#didomi-notice-agree-button").click()

driver.implicitly_wait(5)

natalidad= driver.find_element("css selector", "#tb1").text


Una vez realizado el código para extraer los datos de natalidad de un año, podremos automatizar el proceso en el resto de años. Para ello, primero, tendremos que generar las diferentes URLs correspondientes a los diferentes años de los que queremos información y almacenarlas en una lista. Una vez tengamos la lista, iteraremos por ella, entrando así en cada enlace y aplicando el código genérico creado. Para evitar que tengamos un error que nos detenga el proceso, tendremos en cuenta que no en todas las URLs nos solicitarán que aceptemos cookies y que, de manera excepcional, saltá un pop up que tendremos que cerrar:

url_list = []

for year in range(2017, 2022):
        url_list.append(f"https://datosmacro.expansion.com/demografia/natalidad?anio={year}")

url_list

driver = webdriver.Chrome(ChromeDriverManager().install())
result_list = []

for i in tqdm(url_list[:5]): 
    driver.get(i)
    try:
        sleep(5)
        driver.find_element("css selector", "#didomi-notice-agree-button").click()
        sleep(5)
        natalidad= driver.find_element("css selector", "#tb1").text
        result_list.append(natalidad)
    except:
        try:
            sleep(5)
            driver.find_element("xpath", '//*[@id="socModal"]/div[2]/div/div[1]/button/span').click()
            sleep(5)
            natalidad= driver.find_element("css selector", "#tb1").text
            result_list.append(natalidad)
        except:
            sleep(5)
            natalidad= driver.find_element("css selector", "#tb1").text
            result_list.append(natalidad)
driver.quit()

#result_list

### · Creación de Dataframe

Una vez obtenidos los datos, nuestro siguiente objetivo será la transformación de la lista a dataframe. Haremos una división de la información por año y almacenada en distintos DataFrames que más tarde uniremos. De este último paso tendremos como resultado los siguientes CSVs: `natalidad2017`, `natalidad2018`, `natalidad2019`, `natalidad2020` y `natalidad2021`.

df = pd.DataFrame(result_list)
df

df.shape

#### 2017

a = df.iloc[0][0].split("\n")
a[:5]

nat2017= pd.DataFrame(a)
nat2017

dfn17=nat2017[1:]
dfn17

dfn17.to_csv('natalidad2017.csv')

n17= dfn17[0].str.split("", expand=True)
n17

#columnas= [nat2017[0][0]]
#nat2017.columns= columnas

#columnas=[i.lower() for i in columnas]
#columnas

#### 2018 

b = df.iloc[1][0].split("\n")
b[:5]

nat2018= pd.DataFrame(b)
nat2018

dfn18= nat2018[1:]
dfn18

dfn18.to_csv('natalidad2018.csv')

#### 2019

c = df.iloc[2][0].split("\n")
c[:5]

nat2019= pd.DataFrame(c)
nat2019

dfn19= nat2019[1:]
dfn19

dfn19.to_csv('natalidad2019.csv')

#### 2020

d = df.iloc[3][0].split("\n")
d[:5]

nat2020= pd.DataFrame(d)
nat2020

dfn20= nat2020[1:]
dfn20

dfn20.to_csv('natalidad2020.csv')

#### 2021

f = df.iloc[4][0].split("\n")
f[:5]

nat2021= pd.DataFrame(f)
nat2021

dfn21= nat2021[1:]
dfn21

dfn21.to_csv('natalidad2021.csv')
