==> notebooks/3_Limpieza de datos: Natalidad.ipynb <==
# Most Polluted Countries - Limpieza de Datos

### · Antes de empezar... 

1. Para gestionar los Dataframes:

#Para mi doc de Pipeline:

import sys 
sys.path.append("../SRC") 
import support as sp 


import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import plotly.express as px
from collections import Counter
from fuzzywuzzy import process, fuzz
import pickle

#Regex
import re

###  · Carga de CSVs

Empezamos cargando datos CSV con nuestros datos scrapeados de la natalidad en el periodo 2017/201

df_n17= pd.read_csv("../data/natalidad/natalidad2017.csv" ,  index_col = 0)
df_n17.head(5)

df_n18= pd.read_csv("../data/natalidad/natalidad2018.csv" ,  index_col = 0)
df_n18.head(5)

df_n19= pd.read_csv("../data/natalidad/natalidad2019.csv" ,  index_col = 0)
df_n19.head(5)

df_n20= pd.read_csv("../data/natalidad/natalidad2020.csv" ,  index_col = 0)
df_n20.head(5)

df_n21= pd.read_csv("../data/natalidad/natalidad2021.csv" ,  index_col = 0)
df_n21.head(5)

### · Limpieza Datos Scrapeados 

df_n17

Empezaremos por limpiar el grueso de nuestros datos, discriminando los países. La decisión de hacerlo de esta manera es porque el split de la mayoría de información lo haremos por espacios y en los países no es válido ya que muchos tienen nombre compuesto.

df_n17= df_n17["0"].str.split("\s\[", regex= True, expand=True)
#df_n17

Para hacer la división del grueso de datos, crearemos un diccionario. Aprovecharemos que la estructura de los datos extraídos es fija para realizar el almacenamiento en el mismo. Lo primero es conocer la longitud de los diferentes paquetes de información, para crear los condicionales de los que dependerá este almacenamiento. En aquellos datos que no estén informados, con el objetivo de mantener la estructura, introduciremos un valor nulo. Este diccionario será nuestro puente para los DataFrames finales. 

lon=[]

for i in df_n17[1]:
    lon.append(len(i.split()))

lon[:5]

numeros_unicos = set(lon)

print(numeros_unicos)

#lon.index(7)

Una vez sabemos la longitud de nuestros distintos paquetes de datos, empezamos con la creación del diccionario:

nat2017 = {
    "year": [],
    "born": [],
    "male_born": [],
    "female_born":[],
    "birth_rate": [],
    "fertility_rate":[],
    "variation":[]
            }

data_list = df_n17[1]

for data in data_list:
    
    a = data.split()
    
    if len(a)== 8:
        nat2017["year"].append(str(a[1]))
        nat2017["born"].append(str(a[2]))
        nat2017["male_born"].append(str(a[3]))
        nat2017["female_born"].append(str(a[4]))
        nat2017["birth_rate"].append(str(a[5]))
        nat2017["fertility_rate"].append(str(a[6]))
        nat2017["variation"].append(str(a[7]))
    
    if len(a)== 7:
        nat2017["year"].append(str(a[1]))
        nat2017["born"].append(str(a[2]))
        nat2017["male_born"].append(str(a[3]))
        nat2017["female_born"].append(str(a[4]))
        nat2017["birth_rate"].append(str(a[5]))
        nat2017["fertility_rate"].append(str(a[6]))
        nat2017["variation"].append(np.nan)    
            
    if len(a)== 6:
        nat2017["year"].append(str(a[1]))
        nat2017["born"].append(str(a[2]))
        nat2017["male_born"].append(np.nan)
        nat2017["female_born"].append(np.nan)
        nat2017["birth_rate"].append(str(a[3]))
        nat2017["fertility_rate"].append(str(a[4]))
        nat2017["variation"].append(str(a[5]))
            
    if len(a)== 5:
        nat2017["year"].append(str(a[1]))
        nat2017["born"].append(np.nan)
        nat2017["male_born"].append(np.nan)
        nat2017["female_born"].append(np.nan)
        nat2017["birth_rate"].append(str(a[2]))
        nat2017["fertility_rate"].append(str(a[3]))
        nat2017["variation"].append(str(a[4]))
            

#control
len(nat2017["year"])

Una vez finalizado nuestro diccionario con el grueso de datos, debemos hacer la limpieza de la columna países. La limpieza se basará en eliminar espacios o "y" en los nombres, en su lugar pondremos "_": 

df_n17[0].unique()[:5]

df_n17[0] = df_n17[0].apply(lambda x: str(x).replace(" y ","_"))
df_n17[0].unique()[:5]

df_n17[0] = df_n17[0].apply(lambda x: str(x).replace(" ","_"))
df_n17[0].unique()[:5]

Ya disponemos de los países limpios por lo que están listos para ser unidos al resto del diccionario. Esto nos formará uno de nuestros cinco DataFrames finales. Haremos un diccionario con los países limpios, que después zipearemos con la información de nuestro anterior diccionario:

paises17={
        "country": []
            }

for i in df_n17[0]:
    
    paises17["country"].append(i)

#paises17

union17 = list(zip(paises17["country"], nat2017["year"], nat2017["born"], nat2017["male_born"], nat2017["female_born"], nat2017["birth_rate"], nat2017["fertility_rate"] , nat2017["variation"]))

###### *PUNTO DE CONTROL* 📍
Necesario para reseteo DataFrame

n17= pd.DataFrame(union17)
#n17

#control
#n17.info()
#n17.isnull().sum()

Nuestro siguiente paso será dar nombre a las columnas:

n17.columns

n17.set_axis(["country", "year", "born", "male_born", "female_born", "birth_rate", "fertility_rate", "variation"], axis=1, inplace=True)

El siguiente paso será gestionar los tipos de datos que tenemos. Nos interesa tener la mayor cantidad de datos numéricos posibles:

n17["year"]= n17["year"].astype("int")

n17["born"] = n17["born"].str.replace(".", "")

n17["born"] = pd.to_numeric(n17["born"])

n17["male_born"] = n17["male_born"].str.replace(".","")

n17["male_born"] = pd.to_numeric(n17["male_born"])

n17["female_born"] = n17["female_born"].str.replace(".","")

n17["female_born"] = pd.to_numeric(n17["female_born"])

##### Gestión de porcentaje  

La gestión a realizar del porcentaje es su eliminación para evitar que sean objetos los números. Primero lo que haremos es sustituir las comas, por puntos, así nos permitirá tener un `float`. Acto seguido eliminados finalmente el porcentaje.  

n17["birth_rate"]= n17["birth_rate"].str.replace("‰","")

n17["birth_rate"]= n17["birth_rate"].str.replace(",","")

n17["birth_rate"] = pd.to_numeric(n17["birth_rate"])

n17["fertility_rate"] = n17["fertility_rate"].str.replace("%","")

n17["fertility_rate"] = n17["fertility_rate"].str.replace(",",".")

n17["fertility_rate"]= pd.to_numeric(n17["fertility_rate"])

n17["variation"] = n17["variation"].str.replace("%","")

n17["variation"]= n17["variation"].str.replace("‰","")

n17["variation"] = n17["variation"].str.replace(",",".")

n17["variation"]= pd.to_numeric(n17["variation"])

#n17

#control para saber como va mi transformación de datos
#n17.info()

#### ¡Lo tenemos! ✨

Primer DataFrame limpio y listo para ser concatenado. Vamos a por el resto.  

### 2018

Ahora a por los otros 4 años de información, donde repetiremos el proceso realizado. Para agilizar el proceso, crearemos funciones de limpieza que iremos aplicando a nuestra información:

df_n18

df_n18= df_n18["0"].str.split("\s\[", regex= True, expand=True)
#df_n18

lon18=[]

for i in df_n18[1]:
    lon18.append(len(i.split()))

unicos18 = set(lon18)

print(unicos18)

nat2018 = {
    "year": [],
    "born": [],
    "male_born": [],
    "female_born":[],
    "birth_rate": [],
    "fertility_rate":[],
    "variation":[]
            }

data_list = df_n18[1]

for data in data_list:
    
    a = data.split()
    
    if len(a)== 8:
        nat2018["year"].append(str(a[1]))
        nat2018["born"].append(str(a[2]))
        nat2018["male_born"].append(str(a[3]))
        nat2018["female_born"].append(str(a[4]))
        nat2018["birth_rate"].append(str(a[5]))
        nat2018["fertility_rate"].append(str(a[6]))
        nat2018["variation"].append(str(a[7]))
    
    if len(a)== 7:
        nat2018["year"].append(str(a[1]))
        nat2018["born"].append(str(a[2]))
        nat2018["male_born"].append(str(a[3]))
        nat2018["female_born"].append(str(a[4]))
        nat2018["birth_rate"].append(str(a[5]))
        nat2018["fertility_rate"].append(str(a[6]))
        nat2018["variation"].append(np.nan)    
            
    if len(a)== 6:
        nat2018["year"].append(str(a[1]))
        nat2018["born"].append(str(a[2]))
        nat2018["male_born"].append(np.nan)
        nat2018["female_born"].append(np.nan)
        nat2018["birth_rate"].append(str(a[3]))
        nat2018["fertility_rate"].append(str(a[4]))
        nat2018["variation"].append(str(a[5]))
            
    if len(a)== 5:
        nat2018["year"].append(str(a[1]))
        nat2018["born"].append(np.nan)
        nat2018["male_born"].append(np.nan)
        nat2018["female_born"].append(np.nan)
        nat2018["birth_rate"].append(str(a[2]))
        nat2018["fertility_rate"].append(str(a[3]))
        nat2018["variation"].append(str(a[4]))

#control
len(nat2018["year"])

df_n18[0] = df_n18[0].apply(lambda x: str(x).replace(" y ","_"))
df_n18[0].unique()[:5]

df_n18[0] = df_n18[0].apply(lambda x: str(x).replace(" ","_"))
df_n18[0].unique()[:5]

paises18={
        "country": []
            }

for i in df_n18[0]:
    
    paises18["country"].append(i)

union18 = list(zip(paises18["country"], nat2018["year"], nat2018["born"], nat2018["male_born"], nat2018["female_born"], nat2018["birth_rate"], nat2018["fertility_rate"] , nat2018["variation"]))

#### *PUNTO DE CONTROL* 📍
Necesario para reseteo DataFrame

n18= pd.DataFrame(union18)
#n18

###### Limpieza datos
Nuestra primera función definida será para establecer las cabeceras de nuestras columnas:

def titulos_col(x):
    x.set_axis(["country", "year", "born", "male_born", "female_born", "birth_rate", "fertility_rate", "variation"], axis=1, inplace=True)

titulos_col(n18)

Nuestras siguientes funciones a definir serán las que nos ayudaran a convertir nuestros objetos en números:

n18["year"]= n18["year"].astype("int")

def numeros(x):
    x = x.str.replace(".","")
    x = pd.to_numeric(x)
    return(x)

sp.numeros(n18["born"])

n18["born"] = numeros(n18["born"])

n18["male_born"]= numeros(n18["male_born"])

n18["female_born"]= numeros(n18["female_born"])

Por último haremos una función que limpie nuestros porcentajes, buscando de nuevo que queden en formato número para poder manejarlos:

def porcentajes (x):
    x = x.str.replace("‰","")
    x = x.str.replace("%","")
    x = x.str.replace(",",".")
    x = pd.to_numeric(x)
    return(x)

n18["birth_rate"]= porcentajes(n18["birth_rate"])

n18["fertility_rate"]= porcentajes(n18["fertility_rate"])

n18["variation"]= porcentajes(n18["variation"])

#n18

#control para saber como va la transformación de datos
#n18.info()

Ahora solo tendríamos que seguir aplicando nuestras funciones a nuestros DataFrames para terminar esta limpieza 🧽🪣  

### 2019

df_n19

df_n19= df_n19["0"].str.split("\s\[", regex= True, expand=True)
#df_n19

lon19=[]

for i in df_n19[1]:
    lon19.append(len(i.split()))

unicos19 = set(lon19)

print(unicos19)

nat2019 = {
    "year": [],
    "born": [],
    "male_born": [],
    "female_born":[],
    "birth_rate": [],
    "fertility_rate":[],
    "variation":[]
            }

data_list = df_n19[1]

for data in data_list:
    
    a = data.split()
    
    if len(a)== 8:
        nat2019["year"].append(str(a[1]))
        nat2019["born"].append(str(a[2]))
        nat2019["male_born"].append(str(a[3]))
        nat2019["female_born"].append(str(a[4]))
        nat2019["birth_rate"].append(str(a[5]))
        nat2019["fertility_rate"].append(str(a[6]))
        nat2019["variation"].append(str(a[7]))
    
    if len(a)== 7:
        nat2019["year"].append(str(a[1]))
        nat2019["born"].append(str(a[2]))
        nat2019["male_born"].append(str(a[3]))
        nat2019["female_born"].append(str(a[4]))
        nat2019["birth_rate"].append(str(a[5]))
        nat2019["fertility_rate"].append(str(a[6]))
        nat2019["variation"].append(np.nan)    
            
    if len(a)== 6:
        nat2019["year"].append(str(a[1]))
        nat2019["born"].append(str(a[2]))
        nat2019["male_born"].append(np.nan)
        nat2019["female_born"].append(np.nan)
        nat2019["birth_rate"].append(str(a[3]))
        nat2019["fertility_rate"].append(str(a[4]))
        nat2019["variation"].append(str(a[5]))
            
    if len(a)== 5:
        nat2019["year"].append(str(a[1]))
        nat2019["born"].append(np.nan)
        nat2019["male_born"].append(np.nan)
        nat2019["female_born"].append(np.nan)
        nat2019["birth_rate"].append(str(a[2]))
        nat2019["fertility_rate"].append(str(a[3]))
        nat2019["variation"].append(str(a[4]))

#control
len(nat2019["year"])

df_n19[0] = df_n19[0].apply(lambda x: str(x).replace(" y ","_"))
df_n19[0].unique()[:5]

df_n19[0] = df_n19[0].apply(lambda x: str(x).replace(" ","_"))
df_n19[0].unique()[:5]

paises19={
        "country": []
            }

for i in df_n19[0]:
    
    paises19["country"].append(i)

union19 = list(zip(paises19["country"], nat2019["year"], nat2019["born"], nat2019["male_born"], nat2019["female_born"], nat2019["birth_rate"], nat2019["fertility_rate"] , nat2019["variation"]))

n19= pd.DataFrame(union19)
#n19

##### Limpiamos Datos 

titulos_col(n19)

n19["year"]= n19["year"].astype("int")

n19["born"]= numeros(n19["born"])

n19["male_born"]= numeros(n19["male_born"])

n19["female_born"]= numeros(n19["female_born"])

n19["birth_rate"] = porcentajes(n19["birth_rate"])

n19["fertility_rate"] = porcentajes(n19["fertility_rate"])

n19["variation"] = porcentajes(n19["variation"])

n19

#control de como va mi transformación de datos:
#n19.info()

### 2020

df_n20

df_n20= df_n20["0"].str.split("\s\[", regex= True, expand=True)
df_n20

lon20=[]

for i in df_n20[1]:
    lon20.append(len(i.split()))

unicos20 = set(lon20)

print(unicos20)

nat2020 = {
    "year": [],
    "born": [],
    "male_born": [],
    "female_born":[],
    "birth_rate": [],
    "fertility_rate":[],
    "variation":[]
            }

data_list = df_n20[1]

for data in data_list:
    
    a = data.split()
    
    if len(a)== 8:
        nat2020["year"].append(str(a[1]))
        nat2020["born"].append(str(a[2]))
        nat2020["male_born"].append(str(a[3]))
        nat2020["female_born"].append(str(a[4]))
        nat2020["birth_rate"].append(str(a[5]))
        nat2020["fertility_rate"].append(str(a[6]))
        nat2020["variation"].append(str(a[7]))
    
    if len(a)== 7:
        nat2020["year"].append(str(a[1]))
        nat2020["born"].append(str(a[2]))
        nat2020["male_born"].append(str(a[3]))
        nat2020["female_born"].append(str(a[4]))
        nat2020["birth_rate"].append(str(a[5]))
        nat2020["fertility_rate"].append(str(a[6]))
        nat2020["variation"].append(np.nan)    
            
    if len(a)== 6:
        nat2020["year"].append(str(a[1]))
        nat2020["born"].append(str(a[2]))
        nat2020["male_born"].append(np.nan)
        nat2020["female_born"].append(np.nan)
        nat2020["birth_rate"].append(str(a[3]))
        nat2020["fertility_rate"].append(str(a[4]))
        nat2020["variation"].append(str(a[5]))
            
    if len(a)== 5:
        nat2020["year"].append(str(a[1]))
        nat2020["born"].append(np.nan)
        nat2020["male_born"].append(np.nan)
        nat2020["female_born"].append(np.nan)
        nat2020["birth_rate"].append(str(a[2]))
        nat2020["fertility_rate"].append(str(a[3]))
        nat2020["variation"].append(str(a[4]))

#control
len(nat2020["year"])

df_n20[0] = df_n20[0].apply(lambda x: str(x).replace(" y ","_"))
df_n20[0].unique()[:5]

df_n20[0] = df_n20[0].apply(lambda x: str(x).replace(" ","_"))
df_n20[0].unique()[:5]

paises20={
        "country": []
            }

for i in df_n20[0]:
    
    paises20["country"].append(i)

union20 = list(zip(paises20["country"], nat2020["year"], nat2020["born"], nat2020["male_born"], nat2020["female_born"], nat2020["birth_rate"], nat2020["fertility_rate"] , nat2020["variation"]))

n20= pd.DataFrame(union20)
n20

##### *Limpiamos Datos* 

titulos_col(n20)

n20["year"]= n20["year"].astype("int")

n20["born"] = numeros(n20["born"])

n20["male_born"]= numeros(n20["male_born"])

n20["female_born"]= numeros(n20["female_born"])

n20["birth_rate"]= porcentajes(n20["birth_rate"])

n20["fertility_rate"]= porcentajes(n20["fertility_rate"])

n20["variation"]= porcentajes(n20["variation"])

#n20

#control transformación datos
#n20.info()

### 2021

df_n21

df_n21= df_n21["0"].str.split("\s\[", regex= True, expand=True)
#df_n21

lon21=[]

for i in df_n21[1]:
    lon21.append(len(i.split()))

unicos21 = set(lon21)

print(unicos21)

nat2021 = {
    "year": [],
    "born": [],
    "male_born": [],
    "female_born":[],
    "birth_rate": [],
    "fertility_rate":[],
    "variation":[]
            }

data_list = df_n21[1]

for data in data_list:
    
    a = data.split()
    
    if len(a)== 8:
        nat2021["year"].append(str(a[1]))
        nat2021["born"].append(str(a[2]))
        nat2021["male_born"].append(str(a[3]))
        nat2021["female_born"].append(str(a[4]))
        nat2021["birth_rate"].append(str(a[5]))
        nat2021["fertility_rate"].append(str(a[6]))
        nat2021["variation"].append(str(a[7]))
    
    if len(a)== 7:
        nat2021["year"].append(str(a[1]))
        nat2021["born"].append(str(a[2]))
        nat2021["male_born"].append(str(a[3]))
        nat2021["female_born"].append(str(a[4]))
        nat2021["birth_rate"].append(str(a[5]))
        nat2021["fertility_rate"].append(str(a[6]))
        nat2021["variation"].append(np.nan)    
            
    if len(a)== 6:
        nat2021["year"].append(str(a[1]))
        nat2021["born"].append(str(a[2]))
        nat2021["male_born"].append(np.nan)
        nat2021["female_born"].append(np.nan)
        nat2021["birth_rate"].append(str(a[3]))
        nat2021["fertility_rate"].append(str(a[4]))
        nat2021["variation"].append(str(a[5]))
            
    if len(a)== 5:
        nat2021["year"].append(str(a[1]))
        nat2021["born"].append(np.nan)
        nat2021["male_born"].append(np.nan)
        nat2021["female_born"].append(np.nan)
        nat2021["birth_rate"].append(str(a[2]))
        nat2021["fertility_rate"].append(str(a[3]))
        nat2021["variation"].append(str(a[4]))

#control
len(nat2021["year"])

df_n21[0] = df_n21[0].apply(lambda x: str(x).replace(" y ","_"))
df_n21[0].unique()[:5]

df_n21[0] = df_n21[0].apply(lambda x: str(x).replace(" ","_"))
df_n21[0].unique()[:5]

paises21={
        "country": []
            }

for i in df_n21[0]:
    
    paises21["country"].append(i)

union21 = list(zip(paises21["country"], nat2021["year"], nat2021["born"], nat2021["male_born"], nat2021["female_born"], nat2021["birth_rate"], nat2021["fertility_rate"] , nat2021["variation"]))

n21= pd.DataFrame(union21)
#n21

##### Limpiamos Datos 

titulos_col(n21)

n21["year"]= n21["year"].astype("int")

n21["born"] = numeros(n21["born"])

n21["male_born"]= numeros(n21["male_born"])

n21["female_born"]= numeros(n21["female_born"])

n21["birth_rate"]= porcentajes(n21["birth_rate"])

n21["fertility_rate"]= porcentajes(n21["fertility_rate"])

n21["variation"]= porcentajes(n21["variation"])

#n21

#control para saber como llevo la transformación de datos
#n21.info()

### Concatenación Tablas Natalidad 

Como punto final, concatenaremos las cinco tablas que hemos limpiado para formar una única tabla con los datos de natalidad comprendidos entre 2017/2021:

natalidad_total= pd.concat([n17, n18, n19, n20, n21], axis=0, ignore_index = True)

natalidad_total

#control todos los registros en tabla final
lentotal= len(n21) + len(n20) + len(n19) + len(n18) + len(n17)
lentotal

natalidad_total.fillna("0", inplace= True)
natalidad_total.head(10)

natalidad_total["variation"]= pd.to_numeric(natalidad_total["variation"])

### Traducción Países DataFrame

pip install deep_translator

from deep_translator import GoogleTranslator

traductor= GoogleTranslator(source='es', target='en')

country= list(natalidad_total["country"])

resultado = traductor.translate_batch(country)

natalidad_total["country"]= resultado

Volvemos a trabajar el formato de la columna países ya que al traducir lo hemos perdido.  

natalidad_total["country"]= natalidad_total["country"].apply(lambda x: str(x).replace(" ","_"))

natalidad_total["country"].unique()

def minusculas (x):
    x= [i.lower() for i in x]
    return(x)

natalidad_total["country"]= minusculas(natalidad_total["country"])

#natalidad_total.head(5)

natalidad_total["country"]= natalidad_total["country"].apply(lambda x: str(x).replace("_"," "))
natalidad_total["country"]

natalidad_total["country"]= natalidad_total["country"].apply(lambda x: str(x).title())

Ahora si que si, tenemos nuestros datos de natalidad listos para ser utilizados. Los guardaremos para utilizarlos más adelante.  

natalidad_total.info()

natalidad_total.to_csv('birthrate.csv')
