==> notebooks/4_Limpieza de datos:  AIR QUALITY Countries and Cities.ipynb <==
# Most Polluted Countries - Limpieza de Datos

### · Antes de empezar... 

1. Para gestionar los Dataframes:

#Para mi doc de Pipeline:

import sys 
sys.path.append("../SRC") 
import support as sp 


import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import plotly.express as px
from collections import Counter
from fuzzywuzzy import process, fuzz
import pickle


###  · Carga de CSVs

Empezamos cargando datos CSV procedentes de Kaggle con la información de los países y las ciudades más contaminadas desde 2017 hasta 2021. En ambos los números están almacenados como objects, por lo que será la limpieza que hagamos a continuación. Empezaremos con el CSV de los países más contaminados:

df_pcountries = pd.read_csv("../data/kaggle/AIR QUALITY INDEX- top countries.csv", index_col = 0 )
df_pcountries.head(2)

len(df_pcountries["Country/Region"].unique())

def tipo_numero(x):
    x.replace("-", "0", inplace= True)
    x= pd.to_numeric(x)
    return(x)

df_pcountries["2020"]= tipo_numero(df_pcountries["2020"])

df_pcountries["2019"]= tipo_numero(df_pcountries["2019"])

df_pcountries["2018"]= tipo_numero(df_pcountries["2018"])

df_pcountries["Population"]= df_pcountries["Population"].str.replace("," , "")

df_pcountries["Population"]= pd.to_numeric(df_pcountries["Population"])

df_pcountries.head(2)

df_pcountries.info()

df_pcountries.to_csv('polluted_countries.csv')

A continuación, haremos la limpieza de las ciudades más contaminadas. En primer lugar dividiremos la ciudad, del país:

df_pcities = pd.read_csv("../data/kaggle/AIR QUALITY INDEX (by cities) - IQAir.csv", index_col = 0 )
df_pcities.head(2)

pc= {
    "city": [],
    "country": [],
    }

for data in df_pcities["City"]:
    
    a= data.split(",")
    
    pc["city"].append(str(a[0]))
    pc["country"].append(str(a[1]))

df_pcities["City"]= pc["city"]

df_pcities["country"]= pc["country"]

#df_pcities.head(2)

def minusculas (x):
    x= [i.lower() for i in x]
    return(x)

df_pcities["City"]= minusculas(df_pcities["City"])

df_pcities["country"]= minusculas(df_pcities["country"])

A continuación los objects a floats:

df_pcities["JAN(2021)"]= tipo_numero(df_pcities["JAN(2021)"])

df_pcities["FEB(2021)"]= tipo_numero(df_pcities["FEB(2021)"])

df_pcities["MAR(2021)"]= tipo_numero(df_pcities["MAR(2021)"])

df_pcities["APR(2021)"]= tipo_numero(df_pcities["APR(2021)"])

df_pcities["MAY(2021)"]= tipo_numero(df_pcities["MAY(2021)"])

df_pcities["JUN(2021)"]= tipo_numero(df_pcities["JUN(2021)"])

df_pcities["JUL(2021)"]= tipo_numero(df_pcities["JUL(2021)"])

df_pcities["AUG(2021)"]= tipo_numero(df_pcities["AUG(2021)"])

df_pcities["SEP(2021)"]= tipo_numero(df_pcities["SEP(2021)"])

df_pcities["OCT(2021)"]= tipo_numero(df_pcities["OCT(2021)"])

df_pcities["NOV(2021)"]= tipo_numero(df_pcities["NOV(2021)"])

df_pcities["DEC(2021)"]= tipo_numero(df_pcities["DEC(2021)"])

df_pcities["2020"]= tipo_numero(df_pcities["2020"])

df_pcities["2019"]= tipo_numero(df_pcities["2019"])

df_pcities["2018"]= tipo_numero(df_pcities["2018"])

df_pcities["2017"]= tipo_numero(df_pcities["2017"])

df_pcities.info()

df_pcities.head(2)

df_pcities.to_csv('polluted_cities.csv')
