==> notebooks/5_Limpieza de Datos: Cost of Living.ipynb <==
# Most Polluted Countries - Limpieza de Datos

### · Antes de empezar... 

1. Para gestionar los Dataframes:

#Para mi doc de Pipeline:

import sys 
sys.path.append("../SRC") 
import support as sp 


import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import plotly.express as px
from collections import Counter
from fuzzywuzzy import process, fuzz
import pickle

###  · Carga de CSVs

Empezamos cargando datos CSV procedentes de Kaggle con la información de los países y las ciudades más contaminadas desde 2017 hasta 2021. 

cof17= pd.read_csv("../data/cof/costliving_2017.csv")
cof17.head(5)

cof17["Unnamed: 0"]

col17= ["countries", "cost_living_index"]

cof17= cof17[col17]
cof17

def minusculas (x):
    x= [i.lower() for i in x]
    return(x)

cof17["countries"]= minusculas(cof17["countries"])

year=["2017"]

year17= year*len(cof17["countries"])

cof17["year"]= year17

cof17["year"]= pd.to_numeric(cof17["year"])

#control de información
#cof17.info()
#cof17

Una vez limpio nuestro DataFrame de coste de vida referente al año 2017, procedemos a limpiar el resto de años:

#### 2018 

cof18= pd.read_csv("../data/cof/costliving_2018.csv")
cof18.head(5)

col18= ["countries", "cost_living_index"]
cof18= cof18[col18]

cof18["countries"]= minusculas(cof18["countries"])

year=["2018"]

year18= year*len(cof18["countries"])

cof18["year"]= year18

cof18["year"]= pd.to_numeric(cof18["year"])

#Control de información
#cof18.info()
cof18.head()

#### 2019 

cof19= pd.read_csv("../data/cof/costliving_2019.csv")
cof19.head(5)

col19= ["countries", "cost_living_index"]
cof19= cof19[col19]

cof19["countries"]= minusculas(cof19["countries"])

year=["2019"]

year19= year*len(cof19["countries"])

cof19["year"]= year19

cof19["year"]= pd.to_numeric(cof19["year"])

#Control de información
cof19.info()
cof19.head(5)

#### 2020

cof20= pd.read_csv("../data/cof/costliving_2019.csv")
cof20.head(5)

col20= ["countries", "cost_living_index"]
cof20= cof20[col20]

cof20["countries"]= minusculas(cof20["countries"])

year=["2020"]

year20= year*len(cof20["countries"])

cof20["year"]= year20

cof20["year"]= pd.to_numeric(cof20["year"])

#Control de información
cof20.info()
cof20.head(5)

#### 2021 

cof21= pd.read_csv("../data/cof/costliving_2021.csv")
cof21.head(5)

col21= ["countries", "cost_living_index"]
cof21= cof21[col21]

cof21["countries"]= minusculas(cof21["countries"])

year=["2021"]

year21= year*len(cof21["countries"])

cof21["year"]= year21

cof21["year"]= pd.to_numeric(cof21["year"])

#Control de información
cof21.info()
cof21.head(5)

costevida_total= pd.concat([cof17, cof18, cof19, cof20, cof21], axis=0, ignore_index = True)

costevida_total.to_csv('cost_living.csv')

costevida_total["countries"]

costevida_total.tail(20)

costevida_total.isnull().sum()

costevida_total.info()
